Increase the masking/UTF-8 validation part in bench.py
Previously the benchmark tested messages of 1000 bytes, which is small,
and so if the masking/validation is even moderately fast, it was drowned
out by the Python code.

Increase the message size to make to make this library's part in it more
pronounced.

Also increase the iteration count to nullify interpreter start up time
(when profiling with external tools, e.g. `perf`, `time`).